requirements.txt
langgraph>=0.0.40
pdfplumber>=0.9.0
openai>=1.0.0
python-dotenv>=1.0.0
typing-extensions>=4.0.0